Editing ItemCode Entry with Lock Group

+++ IspDataDb/dbo/Tables/Religion.sql
@@ -1,5 +1,5 @@
 ﻿CREATE TABLE [dbo].[Religion] (
-    [IdNo]            TINYINT        IDENTITY (1, 1) NOT NULL,
+    [IdNo]            SMALLINT       IDENTITY (1, 1) NOT NULL,
     [ReligionCode]    VARCHAR (5)    COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [ReligionName]    VARCHAR (15)   COLLATE SQL_Latin1_General_CP1_CI_AS NOT NULL,
     [ReligionNameAra] NVARCHAR (30)  COLLATE SQL_Latin1_General_CP1_CI_AS NULL,
@@ -7,6 +7,8 @@
     [DateTimeStamp]   ROWVERSION     NULL,
     CONSTRAINT [PK_Religion] PRIMARY KEY CLUSTERED ([IdNo] ASC)
 );
+
+
